[feat]: 네비게이션 이름 수정 및 파일 이동 (#55)
